+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Added
+
+- add the `devBranch` input to the tasks to specify the branch to make the pull request in the `VtexBuildBeta` task.
+
 ## [1.2.0] - 2024-02-14
 
 ## Changed